pyproject.toml
[tool.poetry]
name = "txtmage"
version = "0.1.0"
description = "A django app that makes use of open AI text to image"
authors = ["Andreas Cormack <andreas@passthekeys.co.uk>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
openai = "^1.7.2"
djangorestframework = "^3.14.0"
Markdown = "^3.5.2"
django-filter = "^23.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
